# openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "LeasingApplicationCreatedDto" : {
        "properties" : {
          "applicationId" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationId" ],
        "type" : "object"
      },
...
      "LeasingApplicationPageDto" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/LeasingApplicationSummaryDto"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "items", "page", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
      "LeasingApplicationSummaryDto" : {
        "properties" : {
          "applicationId" : {
            "type" : "string"
          },
          "bikeId" : {
            "type" : "string"
          },
          "bikeModel" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "customerName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationId", "bikeId", "createdAt", "customerName", "status" ],
        "type" : "object"
      },
      "LeasingRequestInput" : {
        "properties" : {
          "age" : {
            "format" : "int32",
            "type" : "integer"
          },
          "bikeId" : {
            "type" : "string"
          },
          "bikeModel" : {
            "type" : "string"
          },
          "customerName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "monthlyNetIncome" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "required" : [ "age", "bikeId", "bikeModel", "customerName", "email", "monthlyNetIncome" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Customer-portal and back-office endpoints for the MiraVelo bike-leasing process. The engine-internal /engine-rest API is intentionally excluded.",
    "title" : "MiraVelo Bike-Leasing API",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/bike-leasing" : {
      "get" : {
        "operationId" : "listLeasingApplications",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LeasingApplicationPageDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "list-leasing-applications-controller" ]
      },
      "post" : {
        "operationId" : "submitLeasingRequest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LeasingRequestInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LeasingApplicationCreatedDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "submit-leasing-request-controller" ]
      }
    },
...
  }
}